Move JournalReader and JournalWriter to common source set with dummy apple implementation
