+++ lrucache/src/commonMain/kotlin/com/mayakapps/lrucache/journal/JournalReader.kt
@@ -1,0 +1,9 @@
+package com.mayakapps.lrucache.journal
+
+import com.mayakapps.lrucache.Closeable
+
+internal expect class JournalReader(path: String) : Closeable {
+    val isCorrupted: Boolean
+
+    fun readFully(): List<JournalOp>
+}
